refactor: move helper funcs to utils
